% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{get_papers}
\alias{get_papers}
\alias{get_papers.regions}
\title{Papers}
\usage{
get_papers(data, region)

\method{get_papers}{regions}(data, region)
}
\arguments{
\item{data}{Dataset as returned by \code{\link{get_regions}}.}

\item{region}{Name of the region \emph{or} index of region in \code{data}.}
}
\description{
Get papers from a certain region.
}
\examples{
regions <- get_regions()
get_papers(regions, "Belgium")
get_papers(regions, 15) # 15th country in the regions data.frame

}
