% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{papers}
\alias{papers}
\title{Newspapers}
\format{A \code{tibble} object with 1834 nodes and 3 edges
\describe{
  \item{\code{newspaper}}{ Newspaper name}
  \item{\code{link}}{ Link to Newspaper}
  \item{\code{region}}{ Newspaper region}
}}
\source{
\url{https://www.newspaperindex.com/}
}
\usage{
papers
}
\description{
Worldwide newspapers.
}
\keyword{datasets}
