% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{user-agent}
\alias{user-agent}
\alias{get_user_agent}
\alias{set_user_agent}
\title{User Agent}
\usage{
get_user_agent()

set_user_agent(agent)
}
\arguments{
\item{agent}{Agent name.}
}
\description{
Get and set \code{user-agent} for subsequent calls.
}
